+++ app/modules/debug/index.php
@@ -5,6 +5,7 @@ use DebugBar\DataCollector\TimeDataCollector;
 use Pagekit\Debug\DataCollector\AuthDataCollector;
 use Pagekit\Debug\DataCollector\DatabaseDataCollector;
 use Pagekit\Debug\DataCollector\EventDataCollector;
+use Pagekit\Debug\DataCollector\ProfileDataCollector;
 use Pagekit\Debug\DataCollector\RoutesDataCollector;
 use Pagekit\Debug\DataCollector\SystemDataCollector;
 use Pagekit\Debug\DebugBar;
@@ -53,6 +54,7 @@ return [
             $app['debugbar']->addCollector(new TimeDataCollector());
             $app['debugbar']->addCollector(new RoutesDataCollector($app['router'], $app['events'], $app['path.cache']));
             $app['debugbar']->addCollector(new EventDataCollector($app['events'], $app['path']));
+            $app['debugbar']->addCollector(new ProfileDataCollector($app['debugbar.storage']));
 
             if (isset($app['auth'])) {
                 $app['debugbar']->addCollector(new AuthDataCollector($app['auth']));
@@ -77,7 +79,7 @@ return [
                     return;
                 }
 
-                $view->data('$debugbar', ['url' => $app['router']->generate('_debugbar', ['id' => $app['debugbar']->getCurrentRequestId()])]);
+                $view->data('$debugbar', ['current' => $app['debugbar']->getCurrentRequestId()]);
                 $view->style('debugbar', 'app/modules/debug/assets/css/debugbar.css');
                 $view->script('debugbar', 'app/modules/debug/app/bundle/debugbar.js', ['vue']);
             }, 50);