+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Events to DebugBar
 - Current Route Info to DebugBar
+- Request Switcher to DebugBar
 
 ## 0.10.1 (Dezember 17, 2015)